EVG-15420 refactor save project settings for section to use multiple data connectors
